+++ src/MediatR.Useful.Behavior/DependencyInjection/ServiceCollectionExtensions.cs
@@ -1,0 +1,14 @@
+﻿using MediatR.Useful.Behavior.Behavior;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace MediatR.Useful.Behavior.DependencyInjection;
+public static class ServiceCollectionExtensions
+{
+
+    public static void AddAllBehaviors(this IServiceCollection services)
+    {
+        services.AddTransient(typeof(IPipelineBehavior<,>), typeof(ValidationBehaviour<,>));
+        services.AddTransient(typeof(IPipelineBehavior<,>), typeof(CachingBehavior<,>));
+        services.AddTransient(typeof(IPipelineBehavior<,>), typeof(CacheInvalidationBehavior<,>));
+    }
+}